move new actors to a separate file
